Create sharepoint connection for downloading data.

Create sharepoint connection for downloading data.

Create sharepoint connection for downloading data.

Public fields

client

A low-level sharepoint client object, which can be used to interact directly with the sharepoint API. This object mostly handles authentication, etc.

Methods

Public methods

sharepoint$new()

sharepoint$download()

sharepoint$folder()

Method new()

Create sharepoint object for downloading data from sharepoint

Usage

sharepoint$new(sharepoint_url, auth = NULL)

Arguments

sharepoint_url

Root URL of sharepoint site to download from

auth

Authentication data passed to the client

Returns

A new sharepoint object

Method download()

Download data from sharepoint

Usage

sharepoint$download(
  sharepoint_path,
  dest = NULL,
  progress = FALSE,
  overwrite = FALSE
)

Arguments

sharepoint_path

Path to the resource within sharepoint

dest

Path to save downloaded data to

progress

Display a progress bar during download?

overwrite

Overwrite existing files?

Returns

Path to saved data

Method folder()

Create a folder object representing a sharepoint folder, with which one can list, download and upload files. See sharepoint_folder for more details.

Usage

sharepoint$folder(site, path, verify = FALSE)

Arguments

site

The name of the sharepoint site (most likely a short string)

path

Relative path within that shared site. It seems that "Shared Documents" is a common path that most likely represents a "Documents" collection when viewed in the sharepoint web interface.

verify

Logical, indicating if the site/path combination is valid (slower but safer).