Download a dataset from sharepoint

Download a dataset from sharepoint

sharepoint_download(
  sharepoint_url,
  sharepoint_path,
  dest = NULL,
  progress = FALSE,
  overwrite = FALSE
)

Arguments

sharepoint_url
The base URL of sharepoint e.g. https://imperiallondon.sharepoint.com

sharepoint_path
The path to the dataset you want to download - this should include any subsites in the url and should be of the form sites/nested/subsites/docs/path/to/document e.g. if you want to get the file at Data/shape files/example.geojson from the site groupA which is in site facultyA the full path would be sites/facultyA/groupA/docs/Data/shape files/example.geojson You should be able to get this if you locate the data you want to download in a browser and click menu on the RHS of the file name which appears on hover -> Copy link and manually edit to get the file path. See vignette for more details.

dest
Path to location you want to save the data. The default save location is a tempfile with the same file extension as the downloaded file.

progress
If TRUE then HTTP requests will print a progress bar

overwrite
if TRUE then the dest will be ovewritten if it exists (otherwise it an error will be thrown)

Value

Path to downloaded data