Interact with sharepoint folders and their files.

Interact with sharepoint folders and their files.

Interact with sharepoint folders and their files.

Public fields

site

Name of the sharepoint site (readonly)

path

Path of the folder (readonly)

Methods

Public methods

sharepoint_folder$new()

sharepoint_folder$files()

sharepoint_folder$folders()

sharepoint_folder$list()

sharepoint_folder$delete()

sharepoint_folder$parent()

sharepoint_folder$folder()

sharepoint_folder$create()

sharepoint_folder$download()

sharepoint_folder$upload()

Method new()

Create sharepoint_folder object to enable listing, creating downloading and uploading files & folders

Usage

sharepoint_folder$new(client, site, path, verify = FALSE)

Arguments

client

A low-level sharepoint client object, which can be used to interact directly with the sharepoint API.

site

The name of the sharepoint site (most likely a short string)

path

Relative path within that shared site. It seems that "Shared Documents" is a common path that most likely represents a "Documents" collection when viewed in the sharepoint web interface.

verify

Logical, indicating if the site/path combination is valid (slower but safer).

Method files()

List all files within the folder

Usage

sharepoint_folder$files(path = NULL)

Arguments

path

Folder relative to this folder, uses this folder if NULL

Method folders()

List all folders within the folder

Usage

sharepoint_folder$folders(path = NULL)

Arguments

path

Folder relative to this folder, uses this folder if NULL

Method list()

List all folders and files within the folder; this is a convenience wrapper around the files and folders methods.

Usage

sharepoint_folder$list(path = NULL)

Arguments

path

Folder relative to this folder, uses this folder if NULL

Method delete()

Delete a folder. Be extremely careful as you could use this to delete an entire sharepoint. Deleted files are sent to the recycle bin, so can be restored with relative ease, but it will still be alarming. There is a mechanism to prevent accidental deletion by declaring a file that exists within the folder.

Usage

sharepoint_folder$delete(path, check)

Arguments

path

The path to delete. Use NULL to delete the current folder.

check

A file (not folder) that exists directly within path, used as a method to verify that you really do want to delete this folder (to prevent things like accidental deletion of the entire sharepoint, for example).

Method parent()

Create an object referring to the parent folder

Usage

sharepoint_folder$parent(verify = FALSE)

Arguments

verify

Verify that the folder exists (which it must really here)

Method folder()

Create an object referring to a child folder

Usage

sharepoint_folder$folder(path, verify = FALSE)

Arguments

path

The name of the folder, relative to this folder

verify

Verify that the folder exists (which it must really here)

Method create()

Create a folder on sharepoint

Usage

sharepoint_folder$create(path)

Arguments

path

Folder relative to this folder

Method download()

Download a file from a folder

Usage

sharepoint_folder$download(
  path,
  dest = NULL,
  progress = FALSE,
  overwrite = FALSE
)

Arguments

path

The name of the path to download, relative to this folder

dest

Path to save downloaded data to. If NULL then a temporary file with the same file extension as path is used. If coderaw() (or any other raw value) then the raw bytes will be returned.

progress

Display httr's progress bar?

overwrite

Overwrite the file if it exists?

Method upload()

Upload a file into a folder

Usage

sharepoint_folder$upload(path, dest = NULL, progress = FALSE)

Arguments

path

The name of the path to upload, absolute, or relative to R's working directory.

dest

Remote path save downloaded data to, relative to this folder. If NULL then the basename of the file is used.

progress

Display httr's progress bar?